+++ inc/menus.php
@@ -6,8 +6,8 @@ if ( ! function_exists( 'agi_navigation_menus' ) ) {
 function agi_navigation_menus() {
 
 	$locations = array(
-		'primary'	=> __( 'Primary Menu - top of page', 'rincon' ),
-		'footer'	=> __( 'Footer Menu - bottom of page', 'rincon' ),
+		'primary'	=> __( 'Primary Menu - top of page', 'sherpa' ),
+		'footer'	=> __( 'Footer Menu - bottom of page', 'sherpa' ),
 	);
 	register_nav_menus( $locations );
 
@@ -17,5 +17,3 @@ function agi_navigation_menus() {
 add_action( 'init', 'agi_navigation_menus' );
 
 }	
-	
-?>