+++ inc/styles.php
@@ -9,16 +9,14 @@ function sherpa_styles() {
 	} else {
 		$main_css_deps = FALSE;
 	}
-	
+
+	wp_register_style( 'font_awesome', '//maxcdn.bootstrapcdn.com/font-awesome/4.6.3/css/font-awesome.min.css', false, '4.6.3', 'all' );
+	wp_enqueue_style( 'font_awesome' );
+
 	wp_register_style( 'main_css', get_stylesheet_directory_uri() . '/css/style.css', $main_css_deps, '1.0', 'all' );
 	wp_enqueue_style( 'main_css' );
-
-	wp_register_style( 'font_awesome', 'https://maxcdn.bootstrapcdn.com/font-awesome/4.6.3/css/font-awesome.min.css', false, '4.6.3', 'all' );
-	wp_enqueue_style( 'font_awesome' );
 
 }
 
 // Hook into the 'wp_enqueue_scripts' action
 add_action( 'wp_enqueue_scripts', 'sherpa_styles' );
-	
-?>