+++ js/plugins.js
@@ -1,2 +1,2 @@
-// @codekit-prepend "bootstrap.min.js";
+// @codekit-prepend "bootstrap-parts.js";
 // @codekit-prepend "vendor/jquery.colorbox.js";